removed features folder (placed all in app folder)
