+++ src/app/(protected)/dashboard/DashboardStats.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 interface DashboardStatsProps {
   title: string;
@@ -8,15 +8,19 @@ interface DashboardStatsProps {
   colorIndex: number;
 }
 
-const DashboardStats: React.FC<DashboardStatsProps> = (
-  { title, icon, value, description, colorIndex },
-) => {
-  const COLORS: string[] = ["primary", "primary"];
+const DashboardStats: React.FC<DashboardStatsProps> = ({
+  title,
+  icon,
+  value,
+  description,
+  colorIndex,
+}) => {
+  const COLORS: string[] = ['primary', 'primary'];
 
   const getDescStyle = () => {
-    if (description.includes("↗︎")) return "font-bold text-green-600";
-    else if (description.includes("↙")) return "font-bold text-red-400";
-    else return "";
+    if (description.includes('↗︎')) return 'font-bold text-green-600';
+    else if (description.includes('↙')) return 'font-bold text-red-400';
+    else return '';
   };
 
   return (
@@ -25,7 +29,7 @@ const DashboardStats: React.FC<DashboardStatsProps> = (
         <div className={`stat-figure `}>{icon}</div>
         <div className="stat-title ">{title}</div>
         <div className={`stat-value mt-2`}>{value}</div>
-        <div className={"stat-desc " + getDescStyle()}>{description}</div>
+        <div className={'stat-desc ' + getDescStyle()}>{description}</div>
       </div>
     </div>
   );
