+++ src/components/features/common/notification-body-right-drawer.tsx
