+++ src/components/features/login/intro-pointers.tsx
@@ -1,4 +1,4 @@
-"use client";
+'use client';
 
 function IntroPointers() {
   return (
@@ -14,14 +14,14 @@ function IntroPointers() {
         ✓ <span className="font-semibold">Light/dark</span> mode toggle
       </p>
       <p className="py-2 ">
-        ✓ <span className="font-semibold">Redux toolkit</span>{" "}
-        and other utility libraries configured
+        ✓ <span className="font-semibold">Redux toolkit</span> and other utility
+        libraries configured
       </p>
       <p className="py-2  ">
         ✓ User-friendly <span className="font-semibold">documentation</span>
       </p>
       <p className="py-2  mb-4">
-        ✓ <span className="font-semibold">Daisy UI</span> components,{" "}
+        ✓ <span className="font-semibold">Daisy UI</span> components,{' '}
         <span className="font-semibold">Tailwind CSS</span> support
       </p>
     </>
