+++ src/containers/confirmation-modal-body.tsx
@@ -1,7 +1,7 @@
-import { CONFIRMATION_MODAL_CLOSE_TYPES } from "@/helper/app-constants";
-import { useAppDispatch } from "@/lib/hooks";
-import { deleteLead } from "@/features/leads/leadSlice";
-import { showNotification } from "@/features/common/headerSlice";
+import { CONFIRMATION_MODAL_CLOSE_TYPES } from '@/helper/app-constants';
+import { useAppDispatch } from '@/lib/hooks';
+import { deleteLead } from '@/app/(protected)/leads/leadSlice';
+import { showNotification } from '@/components/features/common/headerSlice';
 
 interface Props {
   extraObject?: any;
@@ -17,16 +17,14 @@ function ConfirmationModalBody({ extraObject, closeModal }: Props) {
     if (type === CONFIRMATION_MODAL_CLOSE_TYPES.LEAD_DELETE) {
       // positive response, call api or dispatch redux function
       dispatch(deleteLead({ index }));
-      dispatch(showNotification({ message: "Lead Deleted!", status: 1 }));
+      dispatch(showNotification({ message: 'Lead Deleted!', status: 1 }));
     }
     closeModal();
   };
 
   return (
     <>
-      <p className="text-xl mt-8 text-center">
-        {message}
-      </p>
+      <p className="text-xl mt-8 text-center">{message}</p>
 
       <div className="modal-action mt-12">
         <button className="btn btn-outline" onClick={() => closeModal()}>
