+++ src/containers/left-sidebar.tsx
@@ -6,12 +6,12 @@ import routes from '@/helper/sidebar-routes';
 import { usePathname } from 'next/navigation';
 import { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks';
-import { setPageTitle } from '@/features/common/headerSlice';
+import { setPageTitle } from '@/components/features/common/headerSlice';
+import { getUserInfo } from '@/components/features/common/userSlice';
 import { SidebarMenuObj, UserProfile } from '@/helper/types';
 import BookmarkSquareIcon from '@heroicons/react/24/outline/BookmarkSquareIcon';
 import ChevronUpIcon from '@heroicons/react/24/outline/ChevronUpIcon';
 import ArrowUpOnSquareIcon from '@heroicons/react/24/outline/ArrowUpOnSquareIcon';
-import { getUserInfo } from '@/features/common/userSlice';
 import auth from '@/lib/auth';
 import Image from 'next/image';
