+++ src/containers/nav-menu.tsx
@@ -1,12 +1,9 @@
-import React from "react";
+import React from 'react';
 
-import Link from "next/link";
+import Link from 'next/link';
 
 function NavMenu() {
-  return (
-    <>
-    </>
-  );
+  return <></>;
 }
 
 export default NavMenu;
