+++ src/lib/store.ts
@@ -1,10 +1,10 @@
-import headerSlice from "@/features/common/headerSlice";
-import modalSlice from "@/features/common/modalSlice";
-import rightDrawerSlice from "@/features/common/rightDrawerSlice";
-import userSlice from "@/features/common/userSlice";
-import leadSlice from "@/features/leads/leadSlice";
+import headerSlice from '@/components/features/common/headerSlice';
+import modalSlice from '@/components/features/common/modalSlice';
+import rightDrawerSlice from '@/components/features/common/rightDrawerSlice';
+import userSlice from '@/components/features/common/userSlice';
+import leadSlice from '@/app/(protected)/leads/leadSlice';
 
-import { configureStore } from "@reduxjs/toolkit";
+import { configureStore } from '@reduxjs/toolkit';
 
 export const makeStore = () => {
   return configureStore({
@@ -21,5 +21,5 @@ export const makeStore = () => {
 // Infer the type of makeStore
 export type AppStore = ReturnType<typeof makeStore>;
 // Infer the `RootState` and `AppDispatch` types from the store itself
-export type RootState = ReturnType<AppStore["getState"]>;
-export type AppDispatch = AppStore["dispatch"];
+export type RootState = ReturnType<AppStore['getState']>;
+export type AppDispatch = AppStore['dispatch'];
